requirements.txt
numpy
pandas
opencv-python
Pillow
logzero
scikit-image
smart_open
boto3>=1.21.24
redis
urllib3
snowflake-connector-python
pyarrow==5.0.0
fastparquet
celery
pick
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html 
torch==1.10.0+cu113
torchvision==0.11.1+cu113
statistics
scikit-image
python-dotenv
ftfy
regex
tqdm
mediapipe







